Move local function tests that emit to CodeGen test class
